moved code, added handlers for resetting and stopping
The flag being checked now is platforms.<platform_name>.moving/reset

If moving is true, the platform is moving, otherwise it stops where it's at. if reset is true, then a reset_path is created, and the platforms resets back to origin from where it is. Reset starts it's own loop, in which we don't keep checking if reset is true, which means reset only needs to be true for a tick, and the whole thing needs to go all the way back before it starts up again.