Tweak wait, expose values, tweak positions
The waituntil has an optional check to see if the flag exists at all, and if it doesn't, it stops waiting. That way, if cleanup tries to happen while the platform isn't moving, it'll still progress.

Exposed ease and direction for path blocks. Now any easing can be used, not just cubic inout

Path more accurately gets positions. That is to say, if the position you pass is 0,0,0, the floating block that spawns will be at that location, rather than the armor stand. This makes it easier to position the start/end of platforms.