Update block_moving
Fixed !itworks switch that I was using
No longer remove the shulker from the block when carrying. It's not actually needed
Bit of reorginizing defines and stuff
Paths no longer determine. So we get the paths from the temp flag.
Paths needed to be rounded, since sometime they could be odd
play sound at the two thirds part of path_b
Animations are slower to get the full bounce effect from the block
Block sits slightly above pressure plate, 1 to see the plate so that way players don't forget it's sitting on one, and two, because falling blocks that move into the pressure plate like to pop off