Create burning_leaves.dsc
Burning leaves work as such:
Whenever there's a new set of leaves to be burned, add the three relevant lines to the events (on player shoots *_leaves in:area, and the define and inject). The define assumes a cuboid, but the only tag used is blocks_flagged, so you can actually use any AreaObject. You still have to call it cuboid though because that's what the burn task uses. Then it, if the projectile is one fire, gets all the relavant leaves. In this case, thats all leaves in the area that are flagged with burn_group. if the leaf hit is in burn group A, it gets all burn_group:a leaves. Leaves can only be in one burn group. Then, the list is sorted by distance to which block was hit, so the fire spread will spread from there. The leaves are flagged with "ignited", so that they don't get "double ignited".  Then the leaves run through the remove task one at a time, consecuatively, with a small delay between them all. So the first one will wait 4-8 ticks, and the next one will wait 8-16 ticks (4-8 ticks after the first block) and so on. Also, since the fire_spread is not on in dungeons, the leaf blocks are manually removed as well.